+++ src/main/resources/static/db/ddl.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `users` (
-  `users_id`     BIGINT                          NOT NULL AUTO_INCREMENT,
+  `users_id`     INT                             NOT NULL AUTO_INCREMENT,
   `first_name`   VARCHAR(255)                    NOT NULL,
   `last_name`    VARCHAR(255)                    NOT NULL,
   `password`     VARCHAR(255)                    NOT NULL,
